refactor: implement scanner for aws provider
